Добавлены тесты на выполнение заказов

+++ store/main/models/completed_order.py
@@ -8,7 +8,7 @@ class CompletedOrders(db.Model):
     courier_id = db.Column(db.Integer, db.ForeignKey('courier.id'))
     completed_orders = db.Column(db.Integer, default=0)
     last_complete_time = db.Column(db.DateTime)
-    general_complete_seconds = db.Column(db.DECIMAL)
+    general_complete_seconds = db.Column(db.Float)
     region = db.Column(db.Integer)
 
     def update(self, order):
